+++ src/main/java/com/benmu/framework/constant/WXConstant.java
@@ -24,6 +24,7 @@ public class WXConstant {
     public static final String BACK_PAGE_FAILED = "关闭页面失败";
 
     public static final String ACTION_WEEX_REFRESH = "ACTION_WEEX_REFRESH";
+
     /**
      * weex消息中心
      */
@@ -61,7 +62,7 @@ public class WXConstant {
         public static final String EVENT_BROWSERIMG = "EVENT_BROWSERIMG";
         public static final String EVENT_CAMERA = "EVENT_CAMERA";
         public static final String EVENT_CAMERA_UPLOADIMAGE = "EVENT_CAMERA_UPLOADIMAGE";
-        public static final String EVENT_CAMERA_UPLOAD = "EVENT_CAMERA_UPLOAD";
+        public static final String EVENT_CAMERA_PATH = "EVENT_CAMERA_PATH";
         public static final String EVENT_CAMERA_UPLOADSCREENSHOT = "EVENT_CAMERA_UPLOADSCREENSHOT";
         public static final String EVENT_GEOLOCATION_GET = "EVENT_GEOLOCATION_GET";
         public static final String EVENT_MODAL_ALERT = "EVENT_MODAL_ALERT";
@@ -75,8 +76,9 @@ public class WXConstant {
         public static final String EVENT_COPYSTRING = "EVENT_COPYSTRING";
         public static final String EVENT_RELAYTOFRIEND = "EVENT_RELAYTOFRIEND";
         public static final String EVENT_RELAYTOCRICLE = "EVENT_RELAYTOCRICLE";
-        public static final String EVENT_OPENBROWSER="EVENT_OPENBROWSER";
+        public static final String EVENT_OPENBROWSER = "EVENT_OPENBROWSER";
         public static final String EVENT_COMMUNICATION_SMS = "EVENT_COMMUNICATION_SMS";
         public static final String EVENT_COMMUNICATION_CONTACTS = "EVENT_COMMUNICATION_CONTACTS";
+        public static final String EVENT_SET_HOMEPAGE = "EVENT_SET_HOMEPAGE";
     }
 }